initial ros bringup

+++ fod/FodExtension.cpp
@@ -23,6 +23,6 @@ uint32_t getFodZOrder(uint32_t z, bool touched) {
     return z;
 }
 
-uint32_t getFodUsageBits(uint32_t usageBits, bool) {
+uint64_t getFodUsageBits(uint64_t usageBits, bool) {
     return usageBits;
 }
